SLVS-1564 Move IActiveConfigScopeTracker inside Core assembly. (#5784)
